requirements.txt
streamlit
pandas
scikit-learn
google-cloud-bigquery
google-auth
pyarrow